Complete move to new script pulling from gsheet
